Add staggered grid SBP operators presented in the 2017 JCP paper: Energy stable and high-order-accurate finite difference methods on staggered grids.
This commit addresses some issues with the previous version of staggered grid operators. More specifically, the previous
version suffered from inaccuracies due to round off errors floating point computations of some of the coefficients. The
accuracy has also been improved further by reoptimizing for the free parameters in the boundary closure.

It should be mentioned that the there is an alternate version of staggered grid operators presented in a more recent
paper (Compatible staggered and upwind SBP operators). The operators presented in that paper are found in the directory
"staggered-upwind" and they have both smaller truncation errors and spectral radius compared to the ones found in this
commit. There is also an implementation of an 8th order operator. This commit contains only 2nd, 4th, and 6th order
implementations.